+++ src/jetstream.rs
@@ -54,10 +54,7 @@
 //!
 //! nc.create_stream("my_stream")?;
 //!
-//! let consumer: Consumer = nc.create_consumer("my_stream", "my_consumer")?;
-//!
-//! // process a single item, sending an ack if the closure returns `Ok`.
-//! consumer.process
+//! let consumer: nats::jetstream::Consumer = nc.create_consumer("my_stream", "my_consumer")?;
 //!
 //! # Ok(()) }
 //! ```
@@ -72,7 +69,7 @@
 //!
 //! nc.create_stream("my_stream")?;
 //!
-//! let consumer: Consumer = nc.create_consumer("my_stream", ConsumerConfig {
+//! let consumer: nats::jetstream::Consumer = nc.create_consumer("my_stream", ConsumerConfig {
 //!     durable_name: Some("my_consumer".to_string()),
 //!     deliver_subject: Some("my_push_consumer_subject".to_string()),
 //!     ack_policy: AckPolicy::All,
@@ -109,39 +106,73 @@
 //!
 //! let nc = nats::connect("my_server::4222")?;
 //!
-//! // this will create the consumer if it does not exist already
-//! let consumer = Consumer::create_or_open(nc, "my_stream", "existing_or_created_consumer")?;
+//! // this will create the consumer if it does not exist already.
+//! // consumer must be mut because the `process*` methods perform
+//! // message deduplication using an interval tree, which is
+//! // also publicly accessible via the `Consumer`'s `dedupe_window`
+//! // field.
+//! let mut consumer = Consumer::create_or_open(nc, "my_stream", "existing_or_created_consumer")?;
 //!
-//! // wait indefinitely for the message to arrive
+//! // The `Consumer::process` method executes a closure
+//! // on both push- and pull-based consumers, and if
+//! // the closure returns `Ok` then the message is acked.
+//! // If no message is available, it will wait forever
+//! // for one to arrive.
 //! let msg_data_len: usize = consumer.process(|msg| {
 //!     println!("got message {:?}", msg);
-//!     msg.data.len()
+//!     Ok(msg.data.len())
 //! })?;
 //!
-//! // wait until the consumer's `timeout` field for the message to arrive.
-//! // This can be set manually, and has a very low default of 5ms.
+//! // Similar to `Consumer::process` except wait until the
+//! // consumer's `timeout` field for the message to arrive.
+//! // This can and should be set manually, as it has a low
+//! // default of 5ms.
 //! let msg_data_len: usize = consumer.process_timeout(|msg| {
 //!     println!("got message {:?}", msg);
-//!     msg.data.len()
+//!     Ok(msg.data.len())
 //! })?;
 //!
-//! // wait indefinitely for the first message in a batch, then process
-//! // more messages until the configured timeout is expired
+//! // For consumers operating with `AckPolicy::All`, batch
+//! // processing can provide nice throughput optimizations.
+//! // `Consumer::process_batch` will wait indefinitely for
+//! // the first message in a batch, then process
+//! // more messages until the configured timeout is expired.
+//! // It will batch acks if running with `AckPolicy::All`.
+//! // If there is an error with acking, the last item in the
+//! // returned `Vec` will be the io error. Terminates early
+//! // without acking if the closure returns an `Err`, which
+//! // is included in the final element of the `Vec`. If a
+//! // Timeout happens before the batch size is reached, then
+//! // there will be no errors included in the response `Vec`.
 //! let batch_size = 128;
-//! let results: Vec<usize> = consumer.process_batch(batch_size, |msg| {
+//! let results: Vec<std::io::Result<usize>> = consumer.process_batch(batch_size, |msg| {
 //!     println!("got message {:?}", msg);
-//!     msg.data.len()
-//! })?;
+//!     Ok(msg.data.len())
+//! });
+//! let flipped: std::io::Result<Vec<usize>> = results.into_iter().collect();
+//! let sizes: Vec<usize> = flipped?;
+//!
+//! // For lower-level control for use cases that are not
+//! // well-served by the high-level process* methods,
+//! // there are a number of lower level primitives that
+//! // can be used, such as `Consumer::pull` for pull-based
+//! // consumers and `Message::ack` for manually acking things:
+//! let msg = consumer.pull()?;
+//!
+//! // --- process message ---
+//!
+//! // tell the server the message has been processed
+//! msg.ack()?;
 //!
 //! # Ok(()) }
 //! ```
-//!
 
 use std::{
     collections::VecDeque,
     convert::TryFrom,
     fmt::Debug,
     io::{self, Error, ErrorKind},
+    time::Duration,
 };
 
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
@@ -257,7 +288,7 @@ impl NatsClient {
         let cfg: StreamConfig = stream_config.into();
         if cfg.name.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -273,7 +304,7 @@ impl NatsClient {
     pub fn update_stream(&self, cfg: &StreamConfig) -> io::Result<StreamInfo> {
         if cfg.name.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -323,7 +354,7 @@ impl NatsClient {
         let stream: &str = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -349,7 +380,7 @@ impl NatsClient {
         let stream: &str = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -368,7 +399,7 @@ impl NatsClient {
         let stream: &str = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -387,7 +418,7 @@ impl NatsClient {
         let stream: &str = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -411,7 +442,7 @@ impl NatsClient {
         let stream: &str = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -437,7 +468,7 @@ impl NatsClient {
         let stream = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -485,14 +516,14 @@ impl NatsClient {
         let stream = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
         let consumer = consumer.as_ref();
         if consumer.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the consumer name must not be empty",
             ));
         }
@@ -523,7 +554,7 @@ impl NatsClient {
         let stream: &str = stream.as_ref();
         if stream.is_empty() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
+                ErrorKind::InvalidInput,
                 "the stream name must not be empty",
             ));
         }
@@ -549,7 +580,6 @@ impl NatsClient {
         Res: DeserializeOwned,
     {
         let res_msg = self.request(subject, req)?;
-        println!("got response: {:?}", std::str::from_utf8(&res_msg.data));
         let res: ApiResponse<Res> = serde_json::de::from_slice(&res_msg.data)?;
         match res {
             ApiResponse::Ok(stream_info) => Ok(stream_info),
@@ -587,11 +617,11 @@ pub struct Consumer {
     /// out during `process` and `process_batch`. Defaults
     /// to 5ms, which is likely to be far too low for
     /// workloads crossing physical sites.
-    pub timeout: std::time::Duration,
+    pub timeout: Duration,
 
     /// Contains ranges of processed messages that will be
     /// filtered out upon future receipt.
-    pub dedupe_window: RangeTree,
+    pub dedupe_window: IntervalTree,
 }
 
 impl Consumer {
@@ -673,7 +703,7 @@ impl Consumer {
             stream,
             cfg,
             push_subscriber,
-            timeout: std::time::Duration::from_millis(5),
+            timeout: Duration::from_millis(5),
             dedupe_window: Default::default(),
         })
     }
@@ -695,34 +725,42 @@ impl Consumer {
     /// If an error is encountered while subscribing or acking messages
     /// that may have returned `Ok` from the closure, that Ok will be
     /// present in the returned vector but the last item in the vector
-    /// will be the encountered error.
+    /// will be the encountered error. If the consumer's timeout expires
+    /// before the entire batch is processed, there will be no error
+    /// pushed to the returned `Vec`, it will just be shorter than the
+    /// specified batch size.
+    ///
+    /// All messages are deduplicated using the `Consumer`'s built-in
+    /// `dedupe_window` before being fed to the provided closure. If
+    /// a message that has already been processed is received, it will
+    /// be acked and skipped. Errors for acking deduplicated messages
+    /// are not included in the returned `Vec`.
     ///
     /// Requires the `jetstream` feature.
-    #[doc(hidden)]
     pub fn process_batch<R, F: FnMut(&Message) -> io::Result<R>>(
-        &self,
+        &mut self,
         batch_size: usize,
         mut f: F,
     ) -> Vec<io::Result<R>> {
-        if self.cfg.durable_name.is_none() {
-            return vec![Err(Error::new(
-                ErrorKind::InvalidData,
-                "process and process_batch are only usable from \
-                Pull-based Consumers with a durable_name set",
-            ))];
-        }
-
-        let subject = format!(
-            "{}CONSUMER.MSG.NEXT.{}.{}",
-            self.api_prefix(),
-            self.stream,
-            self.cfg.durable_name.as_ref().unwrap()
-        );
-
         let mut _sub_opt = None;
         let responses = if let Some(ps) = self.push_subscriber.as_ref() {
             ps
         } else {
+            if self.cfg.durable_name.is_none() {
+                return vec![Err(Error::new(
+                    ErrorKind::InvalidInput,
+                    "process and process_batch are only usable from \
+                    Pull-based Consumers if there is a durable_name set",
+                ))];
+            }
+
+            let subject = format!(
+                "{}CONSUMER.MSG.NEXT.{}.{}",
+                self.api_prefix(),
+                self.stream,
+                self.cfg.durable_name.as_ref().unwrap()
+            );
+
             let sub =
                 match self.nc.request_multi(&subject, batch_size.to_string()) {
                     Ok(sub) => sub,
@@ -737,22 +775,23 @@ impl Consumer {
         let start = std::time::Instant::now();
 
         let mut received = 0;
-        let mut acked = 0;
 
-        while let Ok(msg) = responses.next_timeout(if received == 0 {
+        while let Ok(next) = responses.next_timeout(if received == 0 {
             // wait "forever" for first message
-            std::time::Duration::new(std::u64::MAX >> 2, 0)
+            Duration::new(std::u64::MAX >> 2, 0)
         } else {
             self.timeout
                 .checked_sub(start.elapsed())
                 .unwrap_or_default()
         }) {
-            let ret = f(&msg);
+            let next_id = next.jetstream_message_info().unwrap().stream_seq;
 
-            if ret.is_err() {
-                rets.push(ret);
-                return rets;
+            if self.dedupe_window.already_processed(next_id) {
+                let _dont_care_about_success = next.ack();
+                continue;
             }
+
+            let ret = f(&next);
 
             let is_err = ret.is_err();
             rets.push(ret);
@@ -762,27 +801,24 @@ impl Consumer {
                 // if our ack policy is `All`, after breaking.
                 break;
             } else if self.cfg.ack_policy == AckPolicy::Explicit {
-                let res = msg.respond(AckKind::Ack);
+                self.dedupe_window.mark_processed(next_id);
+                let res = next.ack();
                 if let Err(e) = res {
-                    rets.truncate(acked);
                     rets.push(Err(e));
-                } else {
-                    acked += 1;
                 }
             }
 
-            last = Some(msg);
+            last = Some(next);
             received += 1;
             if received == batch_size {
                 break;
             }
         }
 
-        if let Some(msg) = last {
+        if let Some(last) = last {
             if self.cfg.ack_policy == AckPolicy::All {
-                let res = msg.respond(AckKind::Ack);
+                let res = last.ack();
                 if let Err(e) = res {
-                    rets.truncate(acked);
                     rets.push(Err(e));
                 }
             }
@@ -796,37 +832,60 @@ impl Consumer {
     ///
     /// Does not ack the processed message if the internal closure returns an `Err`.
     ///
+    /// All messages are deduplicated using the `Consumer`'s built-in
+    /// `dedupe_window` before being fed to the provided closure. If
+    /// a message that has already been processed is received, it will
+    /// be acked and skipped.
+    ///
+    /// Does not return an `Err` if acking the message is unsuccessful,
+    /// but the message is still marked in the dedupe window. If you
+    /// require stronger processing guarantees, you can manually call
+    /// the `double_ack` method of the argument message. If you require
+    /// both the returned `Ok` from the closure and the `Err` from a
+    /// failed ack, use `process_batch` instead.
+    ///
     /// Requires the `jetstream` feature.
-    #[doc(hidden)]
     pub fn process<R, F: Fn(&Message) -> io::Result<R>>(
-        &self,
+        &mut self,
         f: F,
     ) -> io::Result<R> {
-        if self.cfg.durable_name.is_none() {
-            return Err(Error::new(
-                ErrorKind::InvalidData,
-                "process and process_batch are only usable from \
-                Pull-based Consumers with a durable_name set",
-            ));
-        }
+        loop {
+            let next = if let Some(ps) = &self.push_subscriber {
+                ps.next().unwrap()
+            } else {
+                if self.cfg.durable_name.is_none() {
+                    return Err(Error::new(
+                        ErrorKind::InvalidInput,
+                        "process and process_batch are only usable from \
+                Pull-based Consumers if there is a durable_name set",
+                    ));
+                }
 
-        let subject = format!(
-            "{}CONSUMER.MSG.NEXT.{}.{}",
-            self.api_prefix(),
-            self.stream,
-            self.cfg.durable_name.as_ref().unwrap()
-        );
+                let subject = format!(
+                    "{}CONSUMER.MSG.NEXT.{}.{}",
+                    self.api_prefix(),
+                    self.stream,
+                    self.cfg.durable_name.as_ref().unwrap()
+                );
 
-        let next = if let Some(ps) = &self.push_subscriber {
-            ps.next().unwrap()
-        } else {
-            self.nc.request(&subject, AckKind::Ack)?
-        };
-        let ret = f(&next)?;
-        if self.cfg.ack_policy != AckPolicy::None {
-            next.respond(AckKind::Ack)?;
+                self.nc.request(&subject, AckKind::Ack)?
+            };
+
+            let next_id = next.jetstream_message_info().unwrap().stream_seq;
+
+            if self.dedupe_window.already_processed(next_id) {
+                let _dont_care = next.ack();
+                continue;
+            }
+
+            let ret = f(&next)?;
+            if self.cfg.ack_policy != AckPolicy::None {
+                let _dont_care = next.ack();
+            }
+
+            self.dedupe_window.mark_processed(next_id);
+            return Ok(ret);
         }
-        Ok(ret)
     }
 
     /// Process and acknowledge a single message, waiting up to the `Consumer`'s
@@ -834,17 +893,114 @@ impl Consumer {
     ///
     /// Does not ack the processed message if the internal closure returns an `Err`.
     ///
+    /// All messages are deduplicated using the `Consumer`'s built-in
+    /// `dedupe_window` before being fed to the provided closure. If
+    /// a message that has already been processed is received, it will
+    /// be acked and skipped.
+    ///
+    /// Does not return an `Err` if acking the message is unsuccessful,
+    /// but the message is still marked in the dedupe window. If you
+    /// require stronger processing guarantees, you can manually call
+    /// the `double_ack` method of the argument message. If you require
+    /// both the returned `Ok` from the closure and the `Err` from a
+    /// failed ack, use `process_batch` instead.
+    ///
     /// Requires the `jetstream` feature.
-    #[doc(hidden)]
     pub fn process_timeout<R, F: Fn(&Message) -> io::Result<R>>(
-        &self,
+        &mut self,
         f: F,
     ) -> io::Result<R> {
+        loop {
+            let next = if let Some(ps) = &self.push_subscriber {
+                ps.next_timeout(self.timeout)?
+            } else {
+                if self.cfg.durable_name.is_none() {
+                    return Err(Error::new(
+                        ErrorKind::InvalidInput,
+                        "process and process_batch are only usable from \
+                Pull-based Consumers if there is a a durable_name set",
+                    ));
+                }
+
+                let subject = format!(
+                    "{}CONSUMER.MSG.NEXT.{}.{}",
+                    self.api_prefix(),
+                    self.stream,
+                    self.cfg.durable_name.as_ref().unwrap()
+                );
+
+                self.nc.request(&subject, b"")?
+            };
+
+            let next_id = next.jetstream_message_info().unwrap().stream_seq;
+
+            if self.dedupe_window.already_processed(next_id) {
+                self.dedupe_window.mark_processed(next_id);
+                let _dont_care = next.ack();
+                continue;
+            }
+
+            let ret = f(&next)?;
+            if self.cfg.ack_policy != AckPolicy::None {
+                let _dont_care = next.ack();
+            }
+            return Ok(ret);
+        }
+    }
+
+    /// For pull-based consumers (a consumer where `ConsumerConfig.deliver_subject` is `None`)
+    /// this can be used to request a single message, and wait forever for a response.
+    /// If you require specifying the batch size or using a timeout while consuming the
+    /// responses, use the `pull_opt` method below.
+    ///
+    /// This is a lower-level method and does not filter messages through the `Consumer`'s
+    /// built-in `dedupe_window` as the various `process*` methods do.
+    ///
+    /// Requires the `jetstream` feature.
+    pub fn pull(&mut self) -> io::Result<Message> {
+        let ret_opt = self
+            .pull_opt(NextRequest {
+                batch: 1,
+                ..Default::default()
+            })?
+            .next();
+
+        if let Some(ret) = ret_opt {
+            Ok(ret)
+        } else {
+            Err(Error::new(
+                ErrorKind::BrokenPipe,
+                "The nats client is shutting down.",
+            ))
+        }
+    }
+
+    /// For pull-based consumers (a consumer where `ConsumerConfig.deliver_subject` is `None`)
+    /// this can be used to request a configurable number of messages, as well as specify
+    /// how the server will keep track of this batch request over time. See the docs for
+    /// `NextRequest` for more information about the options.
+    ///
+    /// This is a lower-level method and does not filter messages through the `Consumer`'s
+    /// built-in `dedupe_window` as the various `process*` methods do.
+    ///
+    /// Requires the `jetstream` feature.
+    pub fn pull_opt(
+        &mut self,
+        next_request: NextRequest,
+    ) -> io::Result<crate::Subscription> {
         if self.cfg.durable_name.is_none() {
             return Err(Error::new(
-                ErrorKind::InvalidData,
-                "process and process_batch are only usable from \
+                ErrorKind::InvalidInput,
+                "this method is only usable from \
                 Pull-based Consumers with a durable_name set",
+            ));
+        }
+
+        if self.cfg.deliver_subject.is_none() {
+            return Err(Error::new(
+                ErrorKind::InvalidInput,
+                "this method is only usable from \
+                Pull-based Consumers with a deliver_subject set",
             ));
         }
 
@@ -855,16 +1011,8 @@ impl Consumer {
             self.cfg.durable_name.as_ref().unwrap()
         );
 
-        let next = if let Some(ps) = &self.push_subscriber {
-            ps.next_timeout(self.timeout)?
-        } else {
-            self.nc.request(&subject, b"")?
-        };
-        let ret = f(&next)?;
-        if self.cfg.ack_policy != AckPolicy::None {
-            next.respond(AckKind::Ack)?;
-        }
-        Ok(ret)
+        let req = serde_json::ser::to_vec(&next_request).unwrap();
+        self.nc.request_multi(&subject, &req)
     }
 
     fn api_prefix(&self) -> &str {
@@ -872,23 +1020,29 @@ impl Consumer {
     }
 }
 
-/// Records ranges of acknowledged messages for
+/// Records ranges of acknowledged IDs for
 /// low-memory deduplication.
 #[derive(Default)]
-pub struct RangeTree {
+pub struct IntervalTree {
     // stores interval start-end
     inner: std::collections::BTreeMap<u64, u64>,
 }
 
-impl RangeTree {
+impl IntervalTree {
     /// Mark this ID as being processed. Returns `true`
     /// if this ID was not already marked as processed.
     pub fn mark_processed(&mut self, id: u64) -> bool {
+        if self.inner.is_empty() {
+            self.inner.insert(1, 1);
+            return true;
+        }
+
         let (prev_start, prev_end) = self
             .inner
             .range(..=&id)
             .next_back()
-            .map_or((0, 0), |(s, e)| (*s, *e));
+            .map(|(s, e)| (*s, *e))
+            .unwrap();
 
         if (prev_start..=prev_end).contains(&id) {
             // range already includes id
@@ -956,7 +1110,10 @@ mod test {
 
     #[test]
     fn range_tree() {
-        let mut rt = RangeTree {
+        let mut rt = IntervalTree::default();
+        assert!(rt.mark_processed(1));
+
+        let mut rt = IntervalTree {
             inner: vec![(0, 0), (6, 6)].into_iter().collect(),
         };
 
@@ -970,7 +1127,7 @@ mod test {
         assert!(rt.already_processed(6));
         assert!(!rt.already_processed(7));
 
-        let mut rt = RangeTree {
+        let mut rt = IntervalTree {
             inner: vec![(3, 3), (6, 6)].into_iter().collect(),
         };
 
@@ -982,7 +1139,7 @@ mod test {
         assert!(rt.already_processed(6));
         assert!(!rt.already_processed(7));
 
-        let mut rt = RangeTree {
+        let mut rt = IntervalTree {
             inner: vec![(0, 0), (5, 5)].into_iter().collect(),
         };
         rt.mark_processed(4);
@@ -992,7 +1149,7 @@ mod test {
         assert!(rt.already_processed(5));
         assert!(!rt.already_processed(6));
 
-        let mut rt = RangeTree {
+        let mut rt = IntervalTree {
             inner: vec![(2, 3), (5, 6)].into_iter().collect(),
         };
         rt.mark_processed(4);